Move `dbSchemeToLanguage` into QueryLanguage file
